Rating changes pdf done
Now rating changes PDF can be created, this is made by copying the contents of get standings mostly.
from the dashboard the page to get rating changes is also linked.

Problems: code used for searching requied users from the list of all the rating changes is like for every user see if he is in the list, this  is bad and needs to be optimized.